# reduced allocations in Bare for big speedup