@dreadl0ck 2018-11-18
I've updated the benchmarks and added length checks in case the arrays are empty,
to prevent accessing the index -1.

Thanks again for your contribution!